Lemon PB: PoC Folder
In `lemon-pb/` add new `poc/` subfolder for hosting Lemon C examples
that are manually ported to PB in order to obtain proof of concept code
which we can then use as a reference for all tweaks to LemonPB sources
and the `lempar.pb` template.

Add Calculator example in `poc/calc/`, still unported but with some file
names and extensions already adapted to PB files.